offchain/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
asyncmy==0.2.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Email
aiosmtplib==3.0.1

# Utilities
httpx==0.26.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.5

# Logging
python-json-logger==2.0.7
